+++ spec/examples.txt
@@ -1,29 +1,29 @@
 example_id                                              | status  | run_time        |
 ------------------------------------------------------- | ------- | --------------- |
-./spec/controllers/ensemble_controller_spec.rb[1:1:1]   | passed  | 0.00559 seconds |
-./spec/controllers/projects_controller_spec.rb[1:1:1]   | passed  | 0.04552 seconds |
-./spec/controllers/projects_controller_spec.rb[1:2:1:1] | passed  | 0.0208 seconds  |
-./spec/controllers/projects_controller_spec.rb[1:2:1:2] | passed  | 0.0479 seconds  |
-./spec/controllers/projects_controller_spec.rb[1:2:1:3] | passed  | 0.02356 seconds |
-./spec/controllers/projects_controller_spec.rb[1:3:1]   | passed  | 0.03025 seconds |
-./spec/controllers/projects_controller_spec.rb[1:3:2]   | passed  | 0.28757 seconds |
-./spec/controllers/tracks_controller_spec.rb[1:1:1]     | passed  | 0.09115 seconds |
-./spec/controllers/tracks_controller_spec.rb[1:2:1]     | passed  | 0.49078 seconds |
-./spec/controllers/tracks_controller_spec.rb[1:2:2]     | passed  | 0.46799 seconds |
-./spec/controllers/tracks_controller_spec.rb[1:2:3]     | passed  | 0.7445 seconds  |
-./spec/controllers/tracks_controller_spec.rb[1:3:1]     | failed  | 0.35258 seconds |
-./spec/controllers/tracks_controller_spec.rb[1:3:2]     | failed  | 0.40598 seconds |
-./spec/features/create_branch_spec.rb[1:1]              | passed  | 0.21345 seconds |
-./spec/features/create_comment_spec.rb[1:1]             | passed  | 0.22721 seconds |
-./spec/features/create_comment_spec.rb[1:2]             | passed  | 0.28825 seconds |
-./spec/features/project_page_spec.rb[1:1]               | passed  | 0.09762 seconds |
-./spec/features/project_page_spec.rb[1:2]               | passed  | 0.09305 seconds |
-./spec/features/project_page_spec.rb[1:3]               | passed  | 0.13978 seconds |
-./spec/features/project_page_spec.rb[1:4]               | passed  | 0.16644 seconds |
-./spec/features/track_delete_spec.rb[1:1:1]             | failed  | 0.10867 seconds |
-./spec/features/upload_spec.rb[1:1]                     | passed  | 2.15 seconds    |
-./spec/features/upload_spec.rb[1:2]                     | passed  | 0.57157 seconds |
-./spec/features/user_spec.rb[1:1:1]                     | passed  | 0.37019 seconds |
-./spec/features/user_spec.rb[1:1:2]                     | passed  | 0.25247 seconds |
+./spec/controllers/ensemble_controller_spec.rb[1:1:1]   | passed  | 0.0966 seconds  |
+./spec/controllers/projects_controller_spec.rb[1:1:1]   | passed  | 0.02793 seconds |
+./spec/controllers/projects_controller_spec.rb[1:2:1:1] | passed  | 0.01804 seconds |
+./spec/controllers/projects_controller_spec.rb[1:2:1:2] | passed  | 0.02533 seconds |
+./spec/controllers/projects_controller_spec.rb[1:2:1:3] | passed  | 0.01736 seconds |
+./spec/controllers/projects_controller_spec.rb[1:3:1]   | passed  | 0.01959 seconds |
+./spec/controllers/projects_controller_spec.rb[1:3:2]   | passed  | 0.01975 seconds |
+./spec/controllers/tracks_controller_spec.rb[1:1:1]     | passed  | 0.05125 seconds |
+./spec/controllers/tracks_controller_spec.rb[1:2:1]     | passed  | 0.35001 seconds |
+./spec/controllers/tracks_controller_spec.rb[1:2:2]     | passed  | 0.04597 seconds |
+./spec/controllers/tracks_controller_spec.rb[1:2:3]     | passed  | 0.41044 seconds |
+./spec/controllers/tracks_controller_spec.rb[1:3:1]     | passed  | 0.92312 seconds |
+./spec/controllers/tracks_controller_spec.rb[1:3:2]     | passed  | 0.55274 seconds |
+./spec/features/create_branch_spec.rb[1:1]              | passed  | 0.19292 seconds |
+./spec/features/create_comment_spec.rb[1:1]             | passed  | 0.18113 seconds |
+./spec/features/create_comment_spec.rb[1:2]             | passed  | 0.23913 seconds |
+./spec/features/project_page_spec.rb[1:1]               | passed  | 0.0804 seconds  |
+./spec/features/project_page_spec.rb[1:2]               | passed  | 0.10559 seconds |
+./spec/features/project_page_spec.rb[1:3]               | passed  | 0.14627 seconds |
+./spec/features/project_page_spec.rb[1:4]               | passed  | 0.16635 seconds |
+./spec/features/track_delete_spec.rb[1:1:1]             | passed  | 1.07 seconds    |
+./spec/features/upload_spec.rb[1:1]                     | passed  | 0.48969 seconds |
+./spec/features/upload_spec.rb[1:2]                     | passed  | 0.60198 seconds |
+./spec/features/user_spec.rb[1:1:1]                     | passed  | 0.06681 seconds |
+./spec/features/user_spec.rb[1:1:2]                     | passed  | 0.16937 seconds |
 ./spec/helpers/branches_helper_spec.rb[1:1]             | pending | 0.00001 seconds |
 ./spec/models/branch_spec.rb[1:1]                       | pending | 0.00005 seconds |
