--- spec/web_helpers.rb
@@ -1,39 +1,0 @@
-
-
-
-def sign_up
-  visit "/users/sign_up"
-  fill_in "user_email",  with: "tom@gmail.com"
-  fill_in "user_password",  with: "password"
-  fill_in "user_password_confirmation",  with: "password"
-  click_button "Sign up"
-end
-
-def sign_in
-  visit "/users/sign_in"
-  fill_in "user_email",  with: "tom@gmail.com"
-  fill_in "user_password",  with: "password"
-  click_button "Log in"
-end
-
-def sign_out
-  click_link ("Logout")
-end
-
-def create_project
-  sign_up
-  click_button "New Project"
-  fill_in "Title", with: "New project"
-  fill_in "Genre", with: "Funk"
-  fill_in "Description", with: "Great new song"
-  click_button "Create Project"
-end
-
-
-def create_branch
-  create_project
-  fill_in "branch_title", with: "New branch"
-  fill_in "Instrument", with: "Trombone"
-  fill_in "Description", with: "Windy"
-  click_button "Create Branch"
-end
